refactor(kwin-effect): extract daemon-apply slots to plasmazoneseffect/daemon_apply.cpp
Pulls 11 daemon-driven apply slots out of plasmazoneseffect.cpp:
emitNavigationFeedback, slotActivateWindowRequested,
slotMoveSpecificWindowToZoneRequested, slotApplyGeometryRequested,
slotApplyGeometriesBatch, slotRaiseWindowsRequested,
slotSnapAllWindowsRequested, slotPendingRestoresAvailable,
slotWindowFloatingChanged, slotWindowMinimizedChanged,
slotRunningWindowsRequested. Drops the main file from 3110 to 2452 lines.
